pyproject.toml
[tool.poetry]
name = "embedding-models"
version = "0.1.0"
description = ""
authors = ["Togglecorp Dev <dev@togglecorp.com>"]
license = "GNU Affero General Public License v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.15"
langchain-community = "^0.2.14"
langchain-openai = "^0.1.23"
sentence-transformers = "3.0.1"
fastapi = "^0.114.0"
uvicorn = "^0.22.0"
sentry-sdk = "^1.5.8"
huggingface-hub = "^0.24.6"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125
# NOTE: Update in .pre-commit-config.yaml as well
extend-exclude = "(__pycache__|.*snap_test_.*\\.py|.+\\/.+\\/migrations\\/.*)"

[tool.isort]
profile = "black"
multi_line_output = 3
skip = [
    "**/__pycache__",
    "**/snap_test_*.py",
    ".venv/",
    "**/migrations/*.py",
]

[tool.pyright]
exclude = [
    "**/__pycache__",
    "**/snap_test_*.py",
    ".venv/",
    "**/migrations/*.py",
]
reportMissingImports = true
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
